refactor: move validator code to student module
